community[patch]: Release 0.0.2
