+++ libs/community/langchain_community/tools/convert_to_openai.py
@@ -1,4 +1,38 @@
-from langchain_community.tools.render import format_tool_to_openai_function
+from langchain_core.tools import BaseTool
 
-# For backwards compatibility
-__all__ = ["format_tool_to_openai_function"]
+from langchain_community.utils.openai_functions import (
+    FunctionDescription,
+    ToolDescription,
+    convert_pydantic_to_openai_function,
+)
+
+
+def format_tool_to_openai_function(tool: BaseTool) -> FunctionDescription:
+    """Format tool into the OpenAI function API."""
+    if tool.args_schema:
+        return convert_pydantic_to_openai_function(
+            tool.args_schema, name=tool.name, description=tool.description
+        )
+    else:
+        return {
+            "name": tool.name,
+            "description": tool.description,
+            "parameters": {
+                # This is a hack to get around the fact that some tools
+                # do not expose an args_schema, and expect an argument
+                # which is a string.
+                # And Open AI does not support an array type for the
+                # parameters.
+                "properties": {
+                    "__arg1": {"title": "__arg1", "type": "string"},
+                },
+                "required": ["__arg1"],
+                "type": "object",
+            },
+        }
+
+
+def format_tool_to_openai_tool(tool: BaseTool) -> ToolDescription:
+    """Format tool into the OpenAI function API."""
+    function = format_tool_to_openai_function(tool)
+    return {"type": "function", "function": function}
