+++ libs/community/pyproject.toml
@@ -1,6 +1,6 @@
 [tool.poetry]
 name = "langchain-community"
-version = "0.0.1"
+version = "0.0.2"
 description = "Community contributed LangChain integrations."
 authors = []
 license = "MIT"
