+++ libs/langchain/tests/unit_tests/tools/test_render.py
@@ -3,7 +3,7 @@ from typing import List
 import pytest
 from langchain_core.tools import BaseTool, tool
 
-from langchain_community.tools.render import (
+from langchain.tools.render import (
     render_text_description,
     render_text_description_and_args,
 )
